# apps/material_agent_service/openapi.yaml
openapi: 3.1.0
info:
  title: Material Agent Service
  description: REST API for VLM-based material assignment to 3D USD objects.
  version: "0.3.0"
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /artifacts/{session_id}/output:
    get:
      summary: Download output USD with materials applied
      operationId: downloadOutput
      tags: [artifacts]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Flattened USD file
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "404": { description: File not available }
  /artifacts/{session_id}/final-render:
    get:
      summary: Download final render image
      operationId: downloadFinalRender
      tags: [artifacts]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: PNG render image
          content:
            image/png:
              schema: { type: string, format: binary }
        "404": { description: Render not available }
  /artifacts/{session_id}/predictions:
    get:
      summary: Download predictions JSONL
      operationId: downloadPredictions
      tags: [artifacts]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Predictions JSONL file
          content:
            application/x-ndjson:
              schema: { type: string }
        "404": { description: Predictions not available }
  /artifacts/{session_id}/report:
    get:
      summary: View prediction HTML report
      operationId: viewReport
      tags: [artifacts]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: HTML prediction report
          content:
            text/html:
              schema: { type: string }
        "404": { description: Not available }
  /artifacts/{session_id}/optimization-report:
    get:
      summary: View optimization JSON report
      operationId: viewOptimizationReport
      tags: [artifacts]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Optimization report
          content:
            application/json:
              schema: { type: object }
        "404": { description: Not available }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Optional Bearer token (or set MATERIAL_AGENT_TOKEN env var)
security:
  - {}
  - bearerAuth: []